+++ packages/powersync-sdk-common/src/client/sync/stream/AbstractStreamingSyncImplementation.ts
@@ -1,4 +1,5 @@
-import _ from 'lodash';
+import throttle from 'lodash/throttle';
+
 import Logger, { ILogger } from 'js-logger';
 
 import {
@@ -16,7 +17,7 @@ import ndjsonStream from 'can-ndjson-stream';
 import { BucketChecksum, BucketStorageAdapter, Checkpoint } from '../bucket/BucketStorageAdapter';
 import { SyncStatus, SyncStatusOptions } from '../../../db/crud/SyncStatus';
 import { SyncDataBucket } from '../bucket/SyncDataBucket';
-import { BaseObserver, BaseListener } from '../../../utils/BaseObserver';
+import { BaseObserver, BaseListener, Disposable } from '../../../utils/BaseObserver';
 
 export enum LockType {
   CRUD = 'crud',
@@ -46,7 +47,27 @@ export interface AbstractStreamingSyncImplementationOptions {
 }
 
 export interface StreamingSyncImplementationListener extends BaseListener {
-  statusChanged?: (status: SyncStatus) => void;
+  statusChanged?: ((status: SyncStatus) => void) | undefined;
+}
+
+export interface StreamingSyncImplementation extends BaseObserver<StreamingSyncImplementationListener>, Disposable {
+  /**
+   * Connects to the sync service
+   */
+  connect(): Promise<void>;
+  /**
+   * Disconnects from the sync services.
+   * @throws if not connected or if abort is not controlled internally
+   */
+  disconnect(): Promise<void>;
+  getWriteCheckpoint: () => Promise<string>;
+  hasCompletedSync: () => Promise<boolean>;
+  isConnected: boolean;
+  lastSyncedAt?: Date;
+  syncStatus: SyncStatus;
+  triggerCrudUpload: () => void;
+  waitForReady(): Promise<void>;
+  waitForStatus(status: SyncStatusOptions): Promise<void>;
 }
 
 export const DEFAULT_CRUD_UPLOAD_THROTTLE_MS = 1000;
@@ -57,9 +78,14 @@ export const DEFAULT_STREAMING_SYNC_OPTIONS = {
   crudUploadThrottleMs: DEFAULT_CRUD_UPLOAD_THROTTLE_MS
 };
 
-export abstract class AbstractStreamingSyncImplementation extends BaseObserver<StreamingSyncImplementationListener> {
+export abstract class AbstractStreamingSyncImplementation
+  extends BaseObserver<StreamingSyncImplementationListener>
+  implements StreamingSyncImplementation
+{
   protected _lastSyncedAt: Date | null;
   protected options: AbstractStreamingSyncImplementationOptions;
+  protected abortController: AbortController | null;
+  protected crudUpdateListener?: () => void;
 
   syncStatus: SyncStatus;
   triggerCrudUpload: () => void;
@@ -69,14 +95,15 @@ export abstract class AbstractStreamingSyncImplementation extends BaseObserver<S
     this.options = { ...DEFAULT_STREAMING_SYNC_OPTIONS, ...options };
     this.syncStatus = new SyncStatus({
       connected: false,
-      lastSyncedAt: null,
+      lastSyncedAt: undefined,
       dataFlow: {
         uploading: false,
         downloading: false
       }
     });
+    this.abortController = null;
 
-    this.triggerCrudUpload = _.throttle(
+    this.triggerCrudUpload = throttle(
       () => {
         if (!this.syncStatus.connected || this.syncStatus.dataFlowStatus.uploading) {
           return;
@@ -88,23 +115,62 @@ export abstract class AbstractStreamingSyncImplementation extends BaseObserver<S
     );
   }
 
+  async waitForReady() {}
+
+  waitForStatus(status: SyncStatusOptions): Promise<void> {
+    return new Promise((resolve) => {
+      const l = this.registerListener({
+        statusChanged: (updatedStatus) => {
+          /**
+           * Match only the partial status options provided in the
+           * matching status
+           */
+          const matchPartialObject = (compA: object, compB: object) => {
+            return Object.entries(compA).every(([key, value]) => {
+              const comparisonBValue = compB[key];
+              if (typeof value == 'object' && typeof comparisonBValue == 'object') {
+                return matchPartialObject(value, comparisonBValue);
+              }
+              return value == comparisonBValue;
+            });
+          };
+
+          if (matchPartialObject(status, updatedStatus.toJSON())) {
+            resolve();
+            l?.();
+          }
+        }
+      });
+    });
+  }
+
   get lastSyncedAt() {
     const lastSynced = this.syncStatus.lastSyncedAt;
     return lastSynced && new Date(lastSynced);
-  }
-
-  protected get logger() {
-    return this.options.logger!;
   }
 
   get isConnected() {
     return this.syncStatus.connected;
   }
 
+  protected get logger() {
+    return this.options.logger!;
+  }
+
+  async dispose() {
+    this.crudUpdateListener?.();
+    this.crudUpdateListener = undefined;
+  }
+
   abstract obtainLock<T>(lockOptions: LockOptions<T>): Promise<T>;
 
   async hasCompletedSync() {
     return this.options.adapter.hasCompletedSync();
+  }
+
+  async getWriteCheckpoint(): Promise<string> {
+    const response = await this.options.remote.get('/write-checkpoint2.json');
+    return response['data']['write_checkpoint'] as string;
   }
 
   protected async _uploadAllCrud(): Promise<void> {
@@ -154,13 +220,38 @@ export abstract class AbstractStreamingSyncImplementation extends BaseObserver<S
     }
   }
 
-  async getWriteCheckpoint(): Promise<string> {
-    const response = await this.options.remote.get('/write-checkpoint2.json');
-    return response['data']['write_checkpoint'] as string;
+  connect() {
+    this.abortController = new AbortController();
+    this.streamingSync(this.abortController.signal);
+    return this.waitForStatus({ connected: true });
   }
 
+  async disconnect(): Promise<void> {
+    if (!this.abortController) {
+      throw new Error('Disconnect not possible');
+    }
+    this.abortController.abort('Disconnected');
+  }
+
+  /**
+   * @deprecated use [connect instead]
+   */
   async streamingSync(signal?: AbortSignal): Promise<void> {
-    signal?.addEventListener('abort', () => {
+    if (!signal) {
+      this.abortController = new AbortController();
+      signal = this.abortController.signal;
+    }
+
+    /**
+     * Listen for CRUD updates and trigger upstream uploads
+     */
+    this.crudUpdateListener = this.options.adapter.registerListener({
+      crudUpdate: () => this.triggerCrudUpload()
+    });
+
+    signal.addEventListener('abort', () => {
+      this.crudUpdateListener?.();
+      this.crudUpdateListener = undefined;
       this.updateSyncStatus({
         connected: false,
         dataFlow: {
@@ -187,7 +278,7 @@ export abstract class AbstractStreamingSyncImplementation extends BaseObserver<S
     }
   }
 
-  async streamingSyncIteration(signal?: AbortSignal, progress?: () => void): Promise<{ retry?: boolean }> {
+  protected async streamingSyncIteration(signal: AbortSignal, progress?: () => void): Promise<{ retry?: boolean }> {
     return await this.obtainLock({
       type: LockType.SYNC,
       signal,
@@ -206,6 +297,7 @@ export abstract class AbstractStreamingSyncImplementation extends BaseObserver<S
           after: after
         }));
 
+        // These are compared by reference
         let targetCheckpoint: Checkpoint | null = null;
         let validatedCheckpoint: Checkpoint | null = null;
         let appliedCheckpoint: Checkpoint | null = null;
@@ -223,7 +315,7 @@ export abstract class AbstractStreamingSyncImplementation extends BaseObserver<S
           // A connection is active and messages are being received
           if (!this.syncStatus.connected) {
             // There is a connection now
-            _.defer(() => this.triggerCrudUpload());
+            Promise.resolve().then(() => this.triggerCrudUpload());
             this.updateSyncStatus({
               connected: true
             });
@@ -245,7 +337,7 @@ export abstract class AbstractStreamingSyncImplementation extends BaseObserver<S
             await this.options.adapter.setTargetCheckpoint(targetCheckpoint);
           } else if (isStreamingSyncCheckpointComplete(line)) {
             this.logger.debug('Checkpoint complete', targetCheckpoint);
-            const result = await this.options.adapter.syncLocalDatabase(targetCheckpoint);
+            const result = await this.options.adapter.syncLocalDatabase(targetCheckpoint!);
             if (!result.checkpointValid) {
               // This means checksums failed. Start again with a new checkpoint.
               // TODO: better back-off
@@ -256,7 +348,7 @@ export abstract class AbstractStreamingSyncImplementation extends BaseObserver<S
               // Continue waiting.
               // landing here the whole time
             } else {
-              appliedCheckpoint = _.clone(targetCheckpoint);
+              appliedCheckpoint = targetCheckpoint;
               this.logger.debug('validated checkpoint', appliedCheckpoint);
               this.updateSyncStatus({
                 connected: true,
@@ -267,7 +359,7 @@ export abstract class AbstractStreamingSyncImplementation extends BaseObserver<S
               });
             }
 
-            validatedCheckpoint = _.clone(targetCheckpoint);
+            validatedCheckpoint = targetCheckpoint;
           } else if (isStreamingSyncCheckpointDiff(line)) {
             // TODO: It may be faster to just keep track of the diff, instead of the entire checkpoint
             if (targetCheckpoint == null) {
@@ -319,13 +411,13 @@ export abstract class AbstractStreamingSyncImplementation extends BaseObserver<S
           } else {
             this.logger.debug('Sync complete');
 
-            if (_.isEqual(targetCheckpoint, appliedCheckpoint)) {
+            if (targetCheckpoint === appliedCheckpoint) {
               this.updateSyncStatus({
                 connected: true,
                 lastSyncedAt: new Date()
               });
-            } else if (_.isEqual(validatedCheckpoint, targetCheckpoint)) {
-              const result = await this.options.adapter.syncLocalDatabase(targetCheckpoint);
+            } else if (validatedCheckpoint === targetCheckpoint) {
+              const result = await this.options.adapter.syncLocalDatabase(targetCheckpoint!);
               if (!result.checkpointValid) {
                 // This means checksums failed. Start again with a new checkpoint.
                 // TODO: better back-off
@@ -335,7 +427,7 @@ export abstract class AbstractStreamingSyncImplementation extends BaseObserver<S
                 // Checksums valid, but need more data for a consistent checkpoint.
                 // Continue waiting.
               } else {
-                appliedCheckpoint = _.clone(targetCheckpoint);
+                appliedCheckpoint = targetCheckpoint;
                 this.updateSyncStatus({
                   connected: true,
                   lastSyncedAt: new Date(),
@@ -355,7 +447,10 @@ export abstract class AbstractStreamingSyncImplementation extends BaseObserver<S
     });
   }
 
-  async *streamingSyncRequest(req: StreamingSyncRequest, signal: AbortSignal): AsyncGenerator<StreamingSyncLine> {
+  protected async *streamingSyncRequest(
+    req: StreamingSyncRequest,
+    signal?: AbortSignal
+  ): AsyncGenerator<StreamingSyncLine> {
     const body = await this.options.remote.postStreaming('/sync/stream', req, {}, signal);
     const stream = ndjsonStream(body);
     const reader = stream.getReader();
@@ -378,7 +473,10 @@ export abstract class AbstractStreamingSyncImplementation extends BaseObserver<S
     const updatedStatus = new SyncStatus({
       connected: options.connected ?? this.syncStatus.connected,
       lastSyncedAt: options.lastSyncedAt ?? this.syncStatus.lastSyncedAt,
-      dataFlow: _.merge(_.clone(this.syncStatus.dataFlowStatus), options.dataFlow ?? {})
+      dataFlow: {
+        ...this.syncStatus.dataFlowStatus,
+        ...options.dataFlow
+      }
     });
 
     if (!this.syncStatus.isEqual(updatedStatus)) {