+++ packages/powersync-sdk-common/src/db/crud/SyncStatus.ts
@@ -1,5 +1,3 @@
-import _ from 'lodash';
-
 export type SyncDataFlowStatus = Partial<{
   downloading: boolean;
   uploading: boolean;
@@ -49,11 +47,19 @@ export class SyncStatus {
   }
 
   isEqual(status: SyncStatus) {
-    return _.isEqual(this.options, status.options);
+    return JSON.stringify(this.options) == JSON.stringify(status.options);
   }
 
   getMessage() {
-    const { dataFlow } = this.options;
+    const dataFlow = this.dataFlowStatus;
     return `SyncStatus<connected: ${this.connected} lastSyncedAt: ${this.lastSyncedAt}. Downloading: ${dataFlow.downloading}. Uploading: ${dataFlow.uploading}`;
+  }
+
+  toJSON(): SyncStatusOptions {
+    return {
+      connected: this.connected,
+      dataFlow: this.dataFlowStatus,
+      lastSyncedAt: this.lastSyncedAt
+    };
   }
 }