+++ packages/powersync-sdk-web/src/index.ts
@@ -2,5 +2,7 @@ export * from '@journeyapps/powersync-sdk-common';
 export * from './db/PowerSyncDatabase';
 export * from './db/sync/WebRemote';
 export * from './db/sync/WebStreamingSyncImplementation';
+export * from './db/sync/SharedWebStreamingSyncImplementation';
 export * from './db/adapters/wa-sqlite/WASQLiteDBAdapter';
 export * from './db/adapters/wa-sqlite/WASQLitePowerSyncDatabaseOpenFactory';
+export * from './db/adapters/AbstractWebPowerSyncDatabaseOpenFactory';