+++ packages/powersync-sdk-web/tests/crud.test.ts
@@ -10,7 +10,7 @@ import {
 } from '@journeyapps/powersync-sdk-common';
 import { WASQLitePowerSyncDatabaseOpenFactory } from '@journeyapps/powersync-sdk-web';
 import { v4 as uuid } from 'uuid';
-import { testSchema } from './test_schema';
+import { testSchema } from './utils/test-schema';
 
 const testId = '2290de4f-0488-4e50-abed-f8e8eb1d0b42';
 
@@ -163,8 +163,8 @@ describe('CRUD Tests', () => {
     const tx = (await powersync.getNextCrudTransaction())!;
     expect(tx.transactionId).equals(1);
     const expectedCrudEntry = new CrudEntry(1, UpdateType.PUT, 'logs', testId, 1, {
-      level: 'INFO',
-      content: 'test log'
+      content: 'test log',
+      level: 'INFO'
     });
     expect(tx.crud[0].equals(expectedCrudEntry)).true;
   });