+++ demos/react-native-supabase-group-chat/CHANGELOG.md
@@ -1,4 +1,14 @@
-# powerchat
+# react-native-supabase-group-chat
+
+## 1.0.2
+
+### Patch Changes
+
+- Updated dependencies [1aed928]
+- Updated dependencies [aede9e7]
+  - @journeyapps/powersync-sdk-react-native@1.3.0
+  - @journeyapps/powersync-sdk-common@1.3.0
+  - @journeyapps/powersync-react@1.0.6
 
 ## 1.0.1