+++ demos/react-native-supabase-todolist/CHANGELOG.md
@@ -1,4 +1,15 @@
-# powersync-example
+# react-native-supabase-todolist
+
+## 1.0.5
+
+### Patch Changes
+
+- Updated dependencies [1aed928]
+- Updated dependencies [aede9e7]
+  - @journeyapps/powersync-sdk-react-native@1.3.0
+  - @journeyapps/powersync-sdk-common@1.3.0
+  - @journeyapps/powersync-attachments@1.1.1
+  - @journeyapps/powersync-react@1.0.6
 
 ## 1.0.4